Ordered iterators.
Add ordered iterators over HAMT and `HashSet`.  Given two sets with the
same elements and the same hashers, the new iterators enumerate these
sets in the same order.  This is in contrast to existing iterators that
do not guarantee consistent ordering for elements with the same hash
values.  Consistent ordering is achieved by sorting collision nodes on
the fly during iteration.

We use the new iterators to fix the implementation of `Ord`, `PartialOrd`,
and `Hash` for `HashSet` (https://github.com/bodil/im-rs/issues/175).